+++ e2e/s3_fake.go
@@ -61,10 +61,7 @@ func s3ServerEndpoint(t *testing.T, testdir *fs.Dir, loglvl, backend string, tim
 	if enableProxy {
 		splitURL := strings.Split(s3srv.URL, ":")
 		port := splitURL[2]
-		// "http://localhost.proxyman.io" is just a dns which points to localhost
-		// it is needed as some systems are hard to not use proxies for localhost traffic.
-		// another alternative workaround could be to use "http://localhost.charlesproxy.com"
-		proxyEnabledURL := "http://localhost.proxyman.io:" + port
+		proxyEnabledURL := "http://localhost.:" + port
 		return proxyEnabledURL, cleanup
 	}
 	return s3srv.URL, cleanup